Fixed new symlink logic in _zipdir() to support both Python 2 and Python 3.
When implementing new logic to properly handle symlinks, the Python 2
long type was used. Python 3 merged "int" and "long" into "int", so this
logic was throwing a SyntaxError. This change adds a check for the
Python version and uses an "int" in the case of Python 3 and a "long" in
the case of Python 2.

sim - https://issues-pdx.amazon.com/issues/BEANSTALK-6619

cr https://cr.amazon.com/r/4706592/