Checking for existing app versions in application before creating one.
Now a customer will be able to create an application version through the
console or APIs and the eb cli will be able to reconize that the
application version exists and deploy it. There is also better error
handling for when an application version is deleted out of band.

SIM https://sim.amazon.com/issues/BEANSTALK-8382

cr https://cr.amazon.com/r/5328721/