The 'eb config' command doesn't show resource name for ELBv2
We are adding a new type of loadbalancer to our supported resources.
We do not want to show the resource names of this option setting because
it is a service given resource name that will not change for the user.

SIM https://sim.amazon.com/issues/BEANSTALK-8273

cr https://cr.amazon.com/r/5361041/