Removed incompatible python 3 line from Progress Stream for eb local run
The old code was attempting to encode every string to utf-8 which would
covert that into a byte. In python 3 they have made explicit differences
between bytes and str which does not allow for method to be applied to
both as they have in the past. I have removed this UTF-8 encoding as it
shuold be unecesarry as the events that we get are already in str
format.

SIM https://sim.amazon.com/issues/BEANSTALK-10945

cr https://cr.amazon.com/r/6752903/