Fixed eb local run to accept volumes outside of /var/app/current
This was required before because volumes with this prefix were assumed
to be local and as a result needed to change their given path to run
docker locally. I've now added the assumption that if that path does not
match that the volume already exists on the docker image or local
environment and the path does not need to change. Also I added support
for the privileged key name.

SIM https://sim.amazon.com/issues/BEANSTALK-9376

cr https://cr.amazon.com/r/6879764/