Make the .elasticbeanstalk/logs/local directory writable by everyone.
We received a ticket describing an EB customer trying to write MySQL
logs to the .elasticbeanstalk/logs/local directory. However, the 'mysql'
user does not have access to this directory.

It would be convenient from the customer's point of view to have access to
all the logs in one place. Hence we make the change to give write access
to all in .elasticbeanstalk/logs/local/.

In the process, we also ensure that the enclosing directory,
'.elasticbeanstalk/logs' has only read permissions for "group" and "other"
users to prevent world users from having access to the "local" dir.

CR: https://cr.amazon.com/r/7445424/