Undoing Recent CodeCommit region additions.
In commit, 3948a02bf79fcbc4eeff4ca522ab9660c031ff4f, we had added two new regions
ca-central-1 and ap-south-1, for Elastic Beanstalk to integrate with CodeCommit in

The Beanstalk Service is presently not in a position to integrate with these
regions. At the same time, there have occurred other changes in the CLI, which we
must release as soon as possible. So, it requires that we undo the recent
CodeCommit region additions.

cr https://cr.amazon.com/r/7395541/