Fix automated tests that break on Windows.
We found many tests that break on Windows due to the following reasons:
1) shutils.rmtree is unable to delete files in the '.git' directory due to permission issues.
This leaves the 'testDir' directory created during test setup intact, hence interfering with
the running of other tests.
2) Windows expects '\r\n' rather than '\n'
3) File paths following UNIX style.
4) Process creation using Popen not working as expected in tests.

We have fixed all instances of the first two cases, and some instances
of the third case. We shall skip the remaining instances of the third case and
all instances of the fourth case on Windows until we can arrive at an elegant
solution for each of them

cr https://cr.amazon.com/r/7407254/