Fix bug in buildops.py caused by Timeout value hardcoded to 5 minutes.
The CLI polls DescribeEvents for events to determine whether CodeBuild
succeeded in generating a build. However, the timeout is hardcoded to
5 minutes after which the CLI incorrectly infers that CodeBuild has
failed, whereas CodeBuild is likely taking a longer time just because
of the size of the build. The EB service also succeeds in creating
and updating environments.

We fix the bug by substituting the hardcoded timeout value with the
value the customer provides in the buildspec.yml. If such a value is not
found, the CLI defaults to using a 60 minute timeout value.

SIM https://sim.amazon.com/issues/BEANSTALK-12013
cr https://cr.amazon.com/r/7495106/