Fix bug where `eb create/deploy --modules ...` commands fail because root-level .elasticbeanstalk/config.yml is absent.
EB CLI supports the ability to create modules/environments (http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/ebcli-compose.html)
and to maintain them as a group.

The structure of such an application on the Customer's machine would
look as follows after `eb init --modules component-a component-b`:

~/project-name
|-- component-a
|   |-- .elasticbeanstalk
|   `-- env.yaml
`-- component-b
    |-- .elasticbeanstalk
    `-- env.yaml

The expectation is then that `eb create/deploy --modules ...` would
create or deploy new versions of each of the components.

This was broken a few months ago when a check was introduced in
ebcore.py to locate the application root-level `.elasticbeanstalk`
directory. The purpose of this check was to determine whether the
directory hosted an Environment application or a Custom Platform
application.

As a result, the CLI complains that there is no EB application rooted
at `project-name`.

We fix the problem by registering regular environment controllers (i.e.
non-platform controllers) and avoiding the check to determine whether
`project-name` contains a root-level `.elasticbeanstalk` directory ONLY
WHEN the `--modules` argument is passed to beanstalk..

SIM: https://sim.amazon.com/issues/BEANSTALK-11996

cr https://cr.amazon.com/r/7514489/