Filter DescribeEvent responses by version label, environment name and request id if they exist
When multiple workflows are triggered from the CLI (possibly through multiple
interfaces (e.g. multiple computers)) associated with a single EB
environment, it is possible that the success of one event might be
misconstrued as the success of another.

For example, this is presently a problem when multiple CodeBuild buildss are in
process and associated with the same environment.

To rectify this, we filter DescribeEvent responses by version label, environment
name, and request_id if they exist. This will narrow the query range hence leading
to accurate results.

SIM: https://sim.amazon.com/issues/BEANSTALK-12211
CR: https://code.amazon.com/reviews/CR-814146