Increase default timeout when database option is specified
The process of creating an environment ('webserver' or 'worker') with a RDS DB can be particularly time-consuming as a result of which the 10 minute default timeout can be insufficient.

To get around this problem, we increase the timeout to 15 minutes when
the '--database' argument is specified with 'eb create'.

SIM: https://sim.amazon.com/issues/BEANSTALK-12212
CR: https://code.amazon.com/reviews/CR-864687