Replace the usage of presigned SigV4 URLs to perform 'git push' with traditional 'git push <remote> <branch>'
When 'eb deploy' is performed through the CLI with CodeCommit as the code
source, although the deployment is successful, and the 'remote' repository
(i.e. CodeCommit) is updated correctly, 'git status' suggests 'local' and
'remote' are out-of-sync.

This is because when a directory containing a 'git' repository is 'eb init'-ed,
and CodeCommit integration is opted into, a corresponding 'remote' CodeCommit
repository is created. At this time, 'git push <presigned url>' is
executed. The critical thing to note here is that this '<presigned url>'
is set as the upstream location (through the '-u' flag to the 'git push'
command)

Upon a subsequent 'eb deploy', a new '<presigned_url>' is generated and
pushed to. After a successful 'eb deploy' (and, as an extension, a
successful 'git push' to CodeCommit), since the remote that is being tracked
is the URL generated during 'eb init', 'git status' shows the diff
w.r.t. to that URL and hence gives customers the perception that 'local'
is still ahead of 'remote', whereas, in reality, the two are in sync.

The AWS CodeCommit documentation(*), makes absolutely no reference to such
a URL. I have verified that 'git push <remote> <branch>' works perfectly.
Since the EBCLI performs 'git push' on behalf of the customer, there is no
reason to not use this.

Since the logic associated with presigning URLs is not longer required,
remove it.

(*) AWS CodeCommit documentation: http://docs.aws.amazon.com/codecommit/latest/userguide/getting-started.html

SIM: https://sim.amazon.com/issues/BEANSTALK-13096