Allow 'eb platform create' to end polling when AMI and service are in different regions.
'eb platform create' allows specifying a 'source_ami' to use in the
custom_platform.json file. It is required for this AMI to exist in the
same region as the one the platform is desired to be created in by the
customer.

When there is a mismatch between the regions of the AMI and the service,
the service recognizes this and returns an error. The EBCLI does not
look for this message because the success-events polling logic is erroneous;
instead of requesting events based on the PlatformArn, it requests events
based on the version label, thereby missing out on information it needs
to know to stop polling the service. This change rectifies the above
inversion.

SIM: https://sim.amazon.com/issues/BEANSTALK-13277

cr https://code.amazon.com/reviews/CR-1115314