Fix CodeCommit problem that is causing repositories with multiple remotes to fail
The git_remote() method that is used to retrieve the name of the remote
repository retrieves the names of all available remotes. This is a problem
when attempting to 'git push' (through 'eb deploy' or 'eb create') -- A
'git push <remote1>\n<remote2>\n... branch-name' is attempted, which is wrong,
since there is no such remote as '<remote1>\n<remote2>\n...'.

SIM: https://sim.amazon.com/issues/BEANSTALK-14029