Timeouts should return exit with code 4
This restores previous behaviour (as of 3.12.0) in the case of some commands and
enables raising TimeoutErrors in the case of the rest that would
previously instead immediately rescue them.

TimeoutErrors will indeed ultimately be resucued in ebrun.py which is where the
exit code of 4 is assigned.

SIM: https://sim.amazon.com/issues/BEANSTALK-14077