Prevent 'eb deploy --modules ...'  from breaking
Presently, the EBCLI looks for a .elasticbeanstalk directory in the root
level, or above , of the project. Whereas the .elasticbeastalk
directories are initialized during 'eb init' in the modules, one is not
created at the root level of the project.

This change prevents the search for such directories at the root-level
or above for .elasticbeanstalk

SIM: https://sim.amazon.com/issues/BEANSTALK-14117