Fix 'eb init --source ...' related bugs
Problem: The migration steps listed in [1] to move from EBCLI 2.6 to
3.12.3 (the latest version at the time of creating this SIM) do not
work.

Bug 1. If the previous version of the AWSEBCLI [2]  is set up, the value
of the region parameter is read by the new EBCLI but subsequently
overridden with a value that is essentially None when the --region flag
is not passed in.

Bug 2. When the --source flag with a valid CodeCommit repository name
and branch name is passed, EBCLI 3:

-	does not set up a corresponding remote in the .git/config file
- does not create the repository/branch if one doesn't exist
- marks the application as a being CodeCommit-source-controlled in the
  '.elasticbeanstalk/config.yml', which is false

This is a bug because if the customer subsequently 'eb create's/'eb
deploy's, the CLI will look up the 'remote' associated with the CodeCommit
repository name and branch name specified in the
'.elasticbeanstalk/config.yml' file and fail.

Furthermore, 'eb init' (without the arguments/interactive flow) sets up
'remote's accurately. It also creates the repository or/and the branch
if they don't already exist.

SIM: https://sim.amazon.com/issues/BEANSTALK-14187

cr https://code.amazon.com/reviews/CR-1494039