Remove code that magically installs Python package 'docker'/Introduce explicit dependency on 'docker'
pip's latest version, 10, removed the 'get_installed_distributions'
method which the EBCLI relied on.

There are other ways to achieve the same, but these are not good
workarounds:

1. 'pip' might have other methods for this, but it is not meant to be imported.
2. 'easy_install' provides a way to install on behalf of the customer,
but we would need to tie the CLI to a particular version of 'setuptools'
for this. Too restrictive for the customer.
3. checking through a `subprocess` the output of `pip list` for
'docker-py' and 'docker' can be error-prone due to a variety of reasons

Furthermore, 'docker-py' was deprecated a year-and-a-half ago so there
is no need to support it anymore.

SIM: https://sim.amazon.com/issues/BEANSTALK-15120