Upgrade version of colorama to match with that of docker-compose
There is an colorama incompatibility warning message when trying to
install EBCLI 3.14.0 on Windows from pip which is throwing customers off.
As a result, they are downgrading to 3.13.0.

Although this is merely a warning message and customers should be able
to normally use the EBCLI, it appears that the warning message is
deterring them from doing so.

SIM: https://sim.amazon.com/issues/BEANSTALK-15786

cr https://code.amazon.com/reviews/CR-2267203