Remove py2exe code from setup.py and remove dependency on pyyaml
It is unclear as to why the check for 'py2exe' in the command line
arguments was ever added. The usage of the EBCLI with 'py2exe' is not
supported or documented.

Furthermore, 'py2exe' supports only Python 3.3 and Python 3.4 but not
Python 2.7, Python 3.5, or Python 3.6, and has not been under active
development since October 2014. It is also classified as a '4- Beta'
project. As such, it would be a big risk to depend on it.

1. So, this commit removes the 'py2exe' code in 'setup.py'

2. This commit also removes the file, 'generate_msi.py', which is ommitted
from the packaged version of the the awsebcli and as such is useless to
customers and useless in the process of EBCLI development. 'generate_msi.py'
was originally added in the commit dd6164d1ce707029eb6045d6025e0587157046d5
and as part of the review: https://cr.amazon.com/r/3386526/

3. Since none of the remaining code in the EBCLI uses 'pyyaml', this
commit eliminates it as a dependency.

py2exe PyPi repository: https://pypi.org/project/py2exe/#history

SIM: https://sim.amazon.com/issues/BEANSTALK-16007

cr https://code.amazon.com/reviews/CR-2432116