Add pyyaml back as a dependency as the 'yaml' module is defined in it
we cannot remove pyyaml as a dependency, as the yaml module, which is at the core of the EBCLI, is defined in pyyaml. The solution is to add pyyaml back as a dependency and to restrict it to the range >=3.10,<=3.12.

SIM: https://sim.amazon.com/issues/BEANSTALK-16058

cr https://code.amazon.com/reviews/CR-2484458