Partially revert 707a7117cf5339675ce11e63a34d367ee8e38de4
Specifically, rever the change made here:
https://code.amazon.com/packages/EB-CLI/blobs/286f4a273ec6fd3b19d017e213d63e173551cb1e/--/ebcli/operations/platformops.py.

The change was originally made in 707a711 due to the impression that
'platformops.list_custom_platform_versions' returns a dictionary with
attributes including a 'PlatformArn' among other attributes, however,
this appears to not be true. Instead, it appears to be returning a list
of PlatformArns. There does appear to be that lingering bug that in
should 'ignored_states' be non-None, the lookup,
"platform['PlatformStatus']" will fail.

SIM: https://sim.amazon.com/issues/BEANSTALK-16219

cr https://code.amazon.com/reviews/CR-2627383