Add Bash and Powershell scripts to run unit tests inside isolated virtual environments and to create awsebcli artifacts
To get closer to Continuous Delivery, the EBCLI end-to-end tests must
run against every new commit of the master branch automatically.
However, for this, they need to know which commit of the EBCLI to run
against.

This commit includes Bash and Powershell scripts that will:

- create a virtual environment for the EBCLI
- run its tests
- ensure there are no dependency mismatches
- package the EBCLI as an artifact if it is a `master` branch.

The artifacts are later expected to be picked up by the end-to-end
testing framework.

SIM: https://sim.amazon.com/issues/BEANSTALK-16868