Bump botocore version to '>=1.12.1,<1.13'/pass serviceIds
Bump botocore version to >=1.12.1,<1.13 and pass serviceIds to the
custom model definitions in the EBCLI.

Although 'botocore' manages service models for the Python SDKs and tools
built on top of it, the EBCLI has been patching a few of these models
itself. As of 1.12, botocore requires that "serviceId" parameters be
passed. This change passes "serviceId"s to the various model definitions
patched by the EBCLI.

SIM: https://sim.amazon.com/issues/BEANSTALK-16879