Add MANIFEST.in with files essential for complete EBCLI packaging
On some systems (e.g. the Jenkins CI, certain important files, such as
CHANGES.rst are omitted from the EBCLI distribution generated through
"python setup.py sdist". The solution to this problem is to include a
MANIFEST.in that will explicitly specify those files that setuptools may
exclude but which must not be omitted.

It is to be noted, a MANIFEST.in was recently removed in commit 15fcfcf
under the assumption it wasn't necessary. This did not however affect
releases as setuptools did not misbehave on the machine on which the
EBCLI was packaged.

SIM: https://sim.amazon.com/issues/BEANSTALK-16983