Pull asciimatics code into the EBCLI/edit EBCLI dependency list
We'll need the code in asciimatics to create pseudo terminals
that are receptive to kestrokes and hence interactive. We also
need a mechanism whereby a terminal screen can be refreshed
and provides scrolling capabilities. asciimatics provides us
with these capabilities. It is licensed until the Apache 2.0
license.

We cannot introduce a direct dependency on the project because
it inturn depends on other libraries that contain C code which
might not compile on certain computers. However, other 'asciimatics'
dependencies which we require have been added as direct
dependencies of the EBCLI. These are 'future' and 'wcwidth'.
Both of these are licensed under MIT.

It is also worth stating that there is already code presently
in the EBCLI that directly works with "curses", however, this
code is hard to reason about and is hard to test. It is
envisioned that asciimatics code will replace the existing low-
level code for 'eb health', 'eb appversion' and the likes.

The biggest advantage that asciimatics brings is that it
is platform agnostic. Presently, there are two different sets
of code for Windows and Unix-based.

It is also worth noting that the present code that interacts
with 'curses' does not allow handling scrollable output.

SIM: https://sim.amazon.com/issues/BEANSTALK-18055
CR: https://code.amazon.com/reviews/CR-5600520