Verify CNAME is available before offering it as the default
In eb create's interactive flow, customers are prompted for a CNAME.
This CNAME is derived from the environment name, but the availablity of
it isn't checked for until after the customer has selected it.

This commit changes the logic such that the CNAME suggested to the
customer is available to begin with.

SIM: https://sim.amazon.com/issues/BEANSTALK-17433

cr https://code.amazon.com/reviews/CR-3906696