fileoperations._traverse_to_project_root --> fileoperations.ProjectRoot.traverse
moving the logic of traversing to the root of the EBCLI project
to a class gives the ability to maintain state so that the
result of this operation can subsequently be cached and redundant
debug log statements can be avoided.

this could just as well have been done through a module global,
however, using a class for this is perhaps cleaner.

SIM: https://sim.amazon.com/issues/BEANSTALK-17536

cr https://code.amazon.com/reviews/CR-4084894