Recognize branch names with forward slashes in the --source argument
The --source argument to several of the eb commands is of the form
codecommit/<repository>/<branch>, where branchs with '/'s are
rejected. With git workflows, it is common to use '/'s in branch
names, so we must permit it.

SIM: https://sim.amazon.com/issues/BEANSTALK-17681

cr https://code.amazon.com/reviews/CR-4259693