Update urllib3, docker-compose, botocore, six dependencies
It has been brought up in
https://forums.aws.amazon.com/thread.jspa?messageID=883931&#883931
that urllib3 and requests (which is a dependency of docker-compose)
that the awsebcli requires presently are vulnerable.

Since we'd now line to depend on urllib3>=1.24.1,<1.25, we need a
version of botocore which supports urllib3>=1.24.1, which is
botocore>=1.12.29.

pytest requires six>=1.10. We update the EBCLI's requirement on six less
restrictively to >=1.11,<1.12 to accommodate certain future
dependencies.

SIM: https://sim.amazon.com/issues/BEANSTALK-17743

cr https://code.amazon.com/reviews/CR-4319845