Introduce direct dependency on requests
When customers pip-upgrade from a version of the awsebcli
earlier than 3.14.9 to 3.14.9, requests might not be updated
to a version which requires urllib3 >=1.24.1,<1.25.

In order to fix this problem, we should pin the version of
requests to 2.20.1,<2.21.

SIM: https://sim.amazon.com/issues/BEANSTALK-17861

cr https://code.amazon.com/reviews/CR-5019397