Ensure deletion of temporary application versions directory
During `eb create` and `eb deploy`, the source in the PWD is zipped, created and then
an application version ZIP file is created from it and
stored in .elasticbeanstalk/app_versions. The resulting file and enclosing
directory are deleted after a call to CreateApplicationVersion is made.

In some cases (reportedly on Windows), it is possible that the resulting
ZIP file might be read-only, making the deletion of the file impossible.
The `eb` in question will fail. By telling "shutil.rmtree" to
“ignore_errors”, we can ensure that the ZIP is deleted and that the call to
`eb deploy` will function normally.

SIM : https://sim.amazon.com/issues/BEANSTALK-17992

CR: https://code.amazon.com/reviews/CR-5482845/