Ensure env variables are checked during eb init
The cli was assuming that if --profile was not passed
then the eb-cli profile should be assumed. This was not
the correct behavior. Instead, it now checks if any
environment variables exist for the AWS credentials.
If they do then no profile is passed.

SIM : https://sim.amazon.com/issues/BEANSTALK-18072

CR : https://code.amazon.com/reviews/CR-5765528/