Prompt customer for application name when '-p' and '-i' arguments are passed to 'eb init'
In 02f74ed00ccd89420152eb03f9ea05bb16475c72, a bug was introduced so
that when the --platform and --interactive arguments were passed to
"eb init", it would not prompt the customer for an application name.

This commit fixes that behaviour. Note that the change in
02f74ed00ccd89420152eb03f9ea05bb16475c72 has not been released yet.

SIM: https://sim.amazon.com/issues/BEANSTALK-18128
CR: https://code.amazon.com/reviews/CR-5878815