Decrease tags key and value character limits by one
The ebcli is currently out of sync with EC2 in how many
characters are accepted in both key and value for tags.
EC2 respects 127 characters for keys and 255 characters
for values. This commit realigns those limits to be the
same for the ebcli.

SIM : https://sim.amazon.com/issues/BEANSTALK-18433