Add additional spot funtionality
After the most recent release there are a few more spot
related features that we need to integrate. In addition to this,
we are also adding the option for min and max instances
for a given environment.

SIM : https://issues.amazon.com/issues/BEANSTALK-20298

CR : https://code.amazon.com/reviews/CR-16602427