Add min and max instances arguments to create
In order to make the spot option `on-demand-base-capacity`
valuable to our customers we need to allow a
finer control over the min and max instances on a
given environment. As such, we're introducing
the --min-instances and --max-instances arguments for create.
If we excluded this option then a customer would only
be able to use the --scale argument and could not
very well customize their spot configuration as
--scale sets both min and max to the same value.

SIM : https://issues.amazon.com/issues/BEANSTALK-20301

CR : https://code.amazon.com/reviews/CR-16681743