remove all setlocale calls
There is currently a bug where the logs command fails if the logs contain unicode characters. This is due to the locale being set to 'C' in a number of places.

The display module has a number of places where the locale is being set to C. This was a bug fix for an issue that was occuring where some locales would where the display module would fail when trying to read a date string with the datetime.strptime function. This function is no longer used so the setlocale calls can be safely removed.

This commit removes all occurances of the locale being set to 'C'.

cr https://code.amazon.com/reviews/CR-22911016