Remove dated docker/eb local setup code
This commit remvoes unnecessary code to check whether the right version
of docker is present. This was necessary a few years ago when older
versions of docker, incompatible with local development, were still in
use. This commit also assumes the customer has the necessary setup to
launch docker locally on a Windows.