Support 'eb deploy ... --archive <ZIP|dir>'
This commit adds support 'eb deploy ... --archive <ZIP|dir>' so that
customers can point to arbitrary directories/ZIP files to deploy to
existing environments.  In the presence of the '--archive' argument, 'eb
deploy' will ignore the .elasticbeanstalk/config.yml file and operate in
its absence. The '--archive' argument will accept directories or ZIP
files. If the argument is a directory, 'eb deploy' will ZIP the
directory and save it in the '~/.ebartifacts/archives' directory. It
will upload it to S3 and invoke EB::CreateApplicationVersion with the
artifact. Like with the eb-deploy flow without '--archive', it will
proceed to perform an EB::UpdateEnvironment and wait for the update to
complete.